requirements.txt
Babel==2.0
Django==1.9
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.9.0
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.6
git+git://github.com/dyve/django-bootstrap3.git@develop
django-gravatar2==1.3.0
django-simple-captcha==0.4.6
docutils==0.12
pep8==1.6.2
pytz==2015.4
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.8